+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/client/TigerRemoteTrafficDownloader.java
@@ -7,50 +7,110 @@ package de.gematik.test.tiger.proxy.client;
 import de.gematik.rbellogger.RbelLogger;
 import de.gematik.rbellogger.data.RbelElement;
 import de.gematik.rbellogger.util.RbelFileWriterUtils;
+import de.gematik.test.tiger.proxy.data.TigerDownloadedMessageFacet;
+import de.gematik.test.tiger.proxy.data.TracingMessagePairFacet;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.atomic.AtomicReference;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import kong.unirest.HttpResponse;
-import kong.unirest.RequestBodyEntity;
 import kong.unirest.Unirest;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.io.FileUtils;
+import lombok.val;
+import org.apache.commons.lang3.StringUtils;
 
-@RequiredArgsConstructor
+@AllArgsConstructor
 @Slf4j
 public class TigerRemoteTrafficDownloader {
 
     private final TigerRemoteProxyClient tigerRemoteProxyClient;
-    private final AtomicReference<String> lastMessageUuid;
 
     public void execute() {
-        final String rawTraffic = downloadTrafficFromRemote();
-        log.info("Downloaded {} of traffic. Now parsing...",
-            FileUtils.byteCountToDisplaySize(rawTraffic.length()));
+        tigerRemoteProxyClient.getTrafficParserExecutor()
+            .submit(tigerRemoteProxyClient::switchToQueueMode);
 
-        parseDownloadedTraffic(rawTraffic);
-        log.info("Successfully downloaded missed traffic from '{}'. Now {} message cached",
-            getRemoteProxyUrl(), getRbelMessages().size());
+        downloadAllTrafficFromRemote();
 
+        log.info("{}Successfully downloaded missed traffic from '{}'.",
+            tigerRemoteProxyClient.proxyName(), getRemoteProxyUrl());
+
+        tigerRemoteProxyClient.getTrafficParserExecutor()
+            .submit(() -> log.info("{}Successfully downloaded & parsed missed traffic from '{}'. Now {} message cached",
+                tigerRemoteProxyClient.proxyName(), getRemoteProxyUrl(), getRbelMessages().size()));
+
+        tigerRemoteProxyClient.getTrafficParserExecutor()
+            .submit(tigerRemoteProxyClient::switchToExecutorMode);
     }
 
-    private void parseDownloadedTraffic(String rawTraffic) {
+    private void parseTrafficChunk(String rawTraffic) {
         final List<RbelElement> convertedMessages = RbelFileWriterUtils.convertFromRbelFile(
             rawTraffic, getRbelLogger().getRbelConverter());
-
-        lastMessageUuid.compareAndSet(null, convertedMessages.get(convertedMessages.size() - 1).getUuid());
+        final long count = rawTraffic.lines().count();
+        convertedMessages.forEach(msg -> msg.addFacet(new TigerDownloadedMessageFacet()));
+        for (int i = 0; i < convertedMessages.size(); i += 2) {
+            val pairFacet = TracingMessagePairFacet.builder()
+                .response(convertedMessages.get(i + 1))
+                .request(convertedMessages.get(i))
+                .build();
+            convertedMessages.get(i + 1).addFacet(pairFacet);
+        }
+        if (log.isTraceEnabled()) {
+            log.trace(
+                "{}Just parsed another traffic batch of {} lines, got {} messages, expected {} (rest was filtered). Now standing at {} messages overall",
+                tigerRemoteProxyClient.proxyName(), count, convertedMessages.size(), (count + 2) / 3,
+                getRbelMessages().size());
+        }
+        if (!convertedMessages.isEmpty()) {
+            tigerRemoteProxyClient.getLastMessageUuid().set(
+                convertedMessages.get(convertedMessages.size() - 1).getUuid()
+            );
+        }
+        convertedMessages.forEach(tigerRemoteProxyClient::triggerListener);
+        if (log.isTraceEnabled()) {
+            log.trace("{}Parsed traffic, ending with {}", tigerRemoteProxyClient.proxyName(),
+                convertedMessages.stream()
+                    .map(RbelElement::getRawStringContent)
+                    .flatMap(content -> Stream.of(content.split(" ")).skip(1).limit(1))
+                    .filter(httpHeaderString -> httpHeaderString.startsWith("/"))
+                    .collect(Collectors.joining(", ")));
+        }
     }
 
-    private String downloadTrafficFromRemote() {
+    private void downloadAllTrafficFromRemote() {
+        PaginationInfo paginationInfo;
+        int pageNumber = 0;
+        // we make a copy of the last uuid because the traffic parsing will be commenced in parallel, meaning
+        // the tigerRemoteProxyClient.getLastMessageUuid() can shift
+        Optional<String> currentLastUuid = Optional.ofNullable(tigerRemoteProxyClient.getLastMessageUuid().get());
+        final int pageSize = tigerRemoteProxyClient.getTigerProxyConfiguration().getTrafficDownloadPageSize();
+        do {
+            paginationInfo = downloadTrafficPageFromRemoteAndAddToQueue(pageSize, currentLastUuid);
+            pageNumber++;
+            currentLastUuid = Optional.ofNullable(paginationInfo.getLastUuid())
+                .filter(StringUtils::isNotEmpty);
+
+            if (pageNumber > 100) {
+                log.warn("Interrupting traffic-download: Reached 100 downloads! "
+                    + "(Maybe the influx of traffic on the upstream proxy is greater then our downstream-sped?)");
+                return;
+            }
+        } while (paginationInfo.getAvailableMessages() > pageSize);
+    }
+
+    private PaginationInfo downloadTrafficPageFromRemoteAndAddToQueue(int pageSize, Optional<String> currentLastUuid) {
         final String downloadUrl = getRemoteProxyUrl() + "/webui/trafficLog.tgr";
-        log.info("Downloading missed traffic from '{}' (currently cached {} messages)",
-            downloadUrl, getRbelMessages().size());
+        log.debug(
+            "{}Downloading missed traffic from '{}', starting from {}. page-size {} (currently cached {} messages)",
+            tigerRemoteProxyClient.proxyName(), currentLastUuid, downloadUrl, pageSize, getRbelMessages().size());
+
         final Map<String, Object> parameters = new HashMap<>();
-        Optional.ofNullable(lastMessageUuid.get())
+        parameters.put("pageSize", pageSize);
+        currentLastUuid
             .ifPresent(uuid -> parameters.put("lastMsgUuid", uuid));
 
         final HttpResponse<String> response = Unirest.get(downloadUrl)
@@ -60,7 +120,8 @@ public class TigerRemoteTrafficDownloader {
             throw new TigerRemoteProxyClientException(
                 "Error while downloading message from remote '" + downloadUrl + "': " + response.getBody());
         }
-        return response.getBody();
+        tigerRemoteProxyClient.getTrafficParserExecutor().submit(() -> parseTrafficChunk(response.getBody()));
+        return PaginationInfo.of(response);
     }
 
     private RbelLogger getRbelLogger() {
@@ -73,5 +134,27 @@ public class TigerRemoteTrafficDownloader {
 
     private String getRemoteProxyUrl() {
         return tigerRemoteProxyClient.getRemoteProxyUrl();
+    }
+
+    @Data
+    @Builder
+    private static class PaginationInfo {
+
+        private final int availableMessages;
+        private final String lastUuid;
+
+        public static PaginationInfo of(HttpResponse<String> response) {
+            return PaginationInfo.builder()
+                .availableMessages(convertHeaderFieldToInt(response, "available-messages"))
+                .lastUuid(response.getHeaders().getFirst("last-uuid"))
+                .build();
+        }
+
+        private static Integer convertHeaderFieldToInt(HttpResponse<String> response, String key) {
+            return Optional.ofNullable(response.getHeaders().getFirst(key))
+                .filter(StringUtils::isNotEmpty)
+                .map(Integer::parseInt)
+                .orElse(-1);
+        }
     }
 }