+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/data/TigerDownloadedMessageFacet.java
@@ -1,0 +1,16 @@
+/*
+ * ${GEMATIK_COPYRIGHT_STATEMENT}
+ */
+
+package de.gematik.test.tiger.proxy.data;
+
+import de.gematik.rbellogger.data.RbelMultiMap;
+import de.gematik.rbellogger.data.facet.RbelFacet;
+
+public class TigerDownloadedMessageFacet implements RbelFacet {
+
+    @Override
+    public RbelMultiMap getChildElements() {
+        return new RbelMultiMap();
+    }
+}