+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/AbstractTigerProxyTest.java
@@ -62,7 +62,6 @@ public abstract class AbstractTigerProxyTest {
                 .withBody(binaryMessageContent)));
 
         RbelOptions.activateJexlDebugging();
-        Unirest.config().reset();
     }
 
     @AfterAll